recognition/requirements.txt
numpy==1.24.4
tqdm==4.67.1
Pillow==11.0.0
opencv-python==4.11.0.86
dv==1.0.12
PyYAML==6.0.2
scipy==1.15.3